Getting a summary with the basic financials for all companies

Getting a summary with the basic financials for all companies

Usage

yearly_data(years = 2020:2023)

Arguments

years

A sequence of numeric values representing the years.

Value

A dataframe

Examples

yearly_data(years = 2022:2023)
#> --------------------------------------
#> -- Starting data retrieval from EDGAR
#> -- May take ~30sec
#> --------------------------------------
#> -- Getting Revenue
#> -- Getting Operating Income
#> -- Getting Net Income
#> -- Getting Gross Profit
#> -- Tidying up the data
#> taxonomyccpuomdata.cikdata.entityNamedata.locdata.startdata.endyearOperatingIncomeLossGrossProfitrevenuenet_income
#> Got the financials of 6640 companies
#> # A tibble: 7,501 × 19
#>    data.cik data.entityName       ccp    year taxonomy uom   data.loc data.start
#>       <int> <chr>                 <chr> <int> <chr>    <chr> <chr>    <chr>     
#>  1  1084869 1 800 FLOWERS COM INC CY20…  2022 us-gaap  USD   US-NY    2021-06-28
#>  2  1084869 1 800 FLOWERS COM INC CY20…  2023 us-gaap  USD   US-NY    2022-07-04
#>  3  1848898 10X CAPITAL VENTURE … CY20…  2022 us-gaap  USD   US-NY    2022-01-01
#>  4  1848948 10X CAPITAL VENTURE … CY20…  2022 us-gaap  USD   US-NY    2022-01-01
#>  5  1770787 10X GENOMICS, INC.    CY20…  2022 us-gaap  USD   US-CA    2022-01-01
#>  6  1738906 111, INC.             CY20…  2022 us-gaap  USD   CN-      2022-01-01
#>  7  1877461 1606 CORP.            CY20…  2022 us-gaap  USD   US-AZ    2022-01-01
#>  8  1821468 17 EDUCATION & TECHN… CY20…  2022 us-gaap  USD   CN-      2022-01-01
#>  9  1690080 180 LIFE SCIENCES CO… CY20…  2022 us-gaap  USD   US-CA    2022-01-01
#> 10  1599407 1847 HOLDINGS LLC     CY20…  2022 us-gaap  USD   US-NY    2022-01-01
#> # ℹ 7,491 more rows
#> # ℹ 11 more variables: data.end <chr>, OperatingIncomeLoss <dbl>,
#> #   GrossProfit <dbl>, revenue <dbl>, net_income <dbl>, change_NI <dbl>,
#> #   change_R <dbl>, change_OI <dbl>, gross_margin <dbl>,
#> #   operating_margin <dbl>, net_margin <dbl>